refactor(configs): #4 improve configs
- Typescript/Angular: put the important options
of `tsconfig.json`which are generated by the
original nx lib generator to the base config.
Why?
Because `@angular-architects/ddd` seems
to assume those options to be in the base
config.

- Also fixing the following warnings during
the execution of `npx nx run non-prod-storybook:test-storybook`:
```
jest-haste-map: Haste module naming collision: @angular-monorepo/shared/ui-menu
  The following files share their name; please adjust your hasteImpl:
    * <rootDir>/libs/shared/ui-menu/package.json
    * <rootDir>/dist/libs/shared/ui-menu/package.json
```
This solution was found in phind.com using the AI model
GPT-4.