feat(central-backend): add route for backend health check
This commit adds a route for backend health check in the
central-backend application. The route is necessary to
prevent timeouts in targets that depend on the backend.
The route returns a 200 status code with an empty response.

The targets which are depending on this route
are e.g. e2e.
```